Break macros crate into two.
This is to avoid mixing a plugin with normal macros.
